fix: work on upsert; tests passing
